refactor: changed folder structure fix: included application.css
